requirements.txt
# Squad API - Python Dependencies
# Story 0.2: Dependencies e Requirements Management
# All versions pinned for reproducibility (Architecture Decision)

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting & Caching
pyrate-limiter==3.1.1
redis[hiredis]==5.0.1

# Async HTTP Clients
aiohttp==3.9.1
httpx==0.25.2

# Retry & Resilience
tenacity==8.2.3

# LLM Provider SDKs
groq==0.4.1
google-genai==0.2.2
anthropic==0.34.2
openai==1.54.3

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
watchdog==3.0.0

# Observability
prometheus-client==0.19.0

# Database
asyncpg==0.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.20.0

# Development Tools
black==23.12.0
ruff==0.1.8
mypy==1.7.1
